refactor(practice-plugin): split 1733-line orchestrator into DDD modules
Extract three domain-focused files from PracticeViewPlugin.tsx:

- freePractice.helpers.ts (57 lines)
  Pure types (MeasureNoteEntry) and functions (finalizeMeasureNotes)
  with no React dependency — independently testable.

- useFreePractice.ts (443 lines)
  Feature 092 domain hook. Owns all free practice state (isFreePractice,
  freeDisplayNotes, freeMidiRecord, etc.), MIDI subscription, measure
  clock, replay timers, and all free-session handlers (start, stop,
  replay, repractice, back, dismiss).

- useSavedPracticeManager.ts (429 lines)
  Features 056/060/061 domain hook. Owns saved practice CRUD, protected
  practice ids, task config, navigation data loading, and save/delete/
  load handlers.

PracticeViewPlugin.tsx shrinks from 1733 → 963 lines and is now a thin
orchestrator that wires the extracted hooks together with the score-based
practice engine (usePracticeLoop, usePracticeMidi, usePracticeHighlights).

No behaviour changes — pure structural refactoring.
